requirements.txt
django>=4.2
djangorestframework>=3.14
celery>=5.3
redis>=5.0
selenium>=4.15
webdriver-manager>=4.0
Pillow>=10.0
python-dotenv>=1.0
google-genai>=1.0
